# Conflicts:
#	sniffy-integration-tests/sniffy-integration-tests-decrypt-tls-bc-after-sniffy/src/test/java/io/sniffy/tls/DecryptBouncyCastleGoogleTrafficTestHelper.java

+++ sniffy-core/src/main/java/io/sniffy/sql/StatementInvocationHandler.java
@@ -102,7 +102,7 @@ class StatementInvocationHandler<T extends Statement> extends SniffyInvocationHa
 
     protected Object invokeTargetAndRecord(Method method, Object[] args, String sql, boolean isUpdateQuery) throws Throwable {
         long start = System.currentTimeMillis();
-        int rowsUpdated = 0;
+        long rowsUpdated = 0;
         try {
             Sniffy.enterJdbcMethod();
             checkConnectionAllowed(true);
